Fixed - Redisson shutdown exception is thrown during background process of expired Tomcat sessions. #6173

+++ redisson-tomcat/redisson-tomcat-10/src/main/java/org/redisson/tomcat/RedissonSessionManager.java
@@ -226,7 +226,8 @@ public class RedissonSessionManager extends ManagerBase {
     public void remove(Session session, boolean update) {
         super.remove(session, update);
         
-        if (session.getIdInternal() != null) {
+        if (session.getIdInternal() != null
+                && !redisson.isShuttingDown()) {
             ((RedissonSession)session).delete();
         }
     }
